# api/docs/swagger/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para a plataforma Traceo — localização de pessoas desaparecidas no Brasil.",
        "title": "Traceo API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/missing": {
            "get": {
                "description": "Retorna lista paginada de desaparecidos (cursor-based)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missing"
                ],
                "summary": "Listar desaparecidos",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Tamanho da página",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor para próxima página",
                        "name": "after",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.MissingListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra uma nova pessoa desaparecida",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missing"
                ],
                "summary": "Cadastrar desaparecido",
                "parameters": [
                    {
                        "description": "Dados do desaparecido",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CreateMissingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.MissingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/missing/locations": {
            "get": {
                "description": "Retorna coordenadas para exibição no mapa",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missing"
                ],
                "summary": "Localizações de desaparecidos",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Limite",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.LocationsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/missing/search": {
            "get": {
                "description": "Busca por prefixo no nome",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missing"
                ],
                "summary": "Buscar desaparecidos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Termo de busca",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limite de resultados",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handler.MissingResponse"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/missing/{id}": {
            "get": {
                "description": "Retorna dados de um desaparecido",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missing"
                ],
                "summary": "Buscar desaparecido por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do desaparecido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.MissingResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza dados de um desaparecido (somente o dono)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missing"
                ],
                "summary": "Atualizar desaparecido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do desaparecido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.UpdateMissingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.MissingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um desaparecido (somente o dono)",
                "tags": [
                    "missing"
                ],
                "summary": "Deletar desaparecido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do desaparecido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sem conteúdo"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/missing/{id}/status": {
            "patch": {
                "description": "Marca como encontrado ou reativa busca",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missing"
                ],
                "summary": "Alterar status do desaparecido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Missing ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{id}/missing": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todos os desaparecidos cadastrados por um usuário",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missing"
                ],
                "summary": "Listar desaparecidos de um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handler.MissingResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.CreateMissingRequest": {
            "type": "object",
            "required": [
                "eyes",
                "gender",
                "hair",
                "name",
                "skin"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "clothes": {
                    "type": "string",
                    "maxLength": 500
                },
                "date_of_disappearance": {
                    "type": "string"
                },
                "event_report": {
                    "type": "string",
                    "maxLength": 2000
                },
                "eyes": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "hair": {
                    "type": "string"
                },
                "height": {
                    "type": "string",
                    "maxLength": 20
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 100
                },
                "photo_url": {
                    "type": "string"
                },
                "scar_description": {
                    "type": "string",
                    "maxLength": 500
                },
                "skin": {
                    "type": "string"
                },
                "tattoo_description": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "internal_handler.MissingListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler.MissingResponse"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.UpdateStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}